[GEOT-7517] Cannot render rasters reprojecting from WGS84 to orthographic projection
